First crude implementation of FTDI MPSSE debug probes
Code structure based on the picoprobe built-in module, with the USB
protocol layer heavily based off of the mpsse access layer in OpenOCD.

All accesses are synchronous, so performance is not great.

Only SWD protocol is implemented so far.

Currently hardcoded to a single USB VID/PID, should be configurable.